Redid raw image uuids and added initial visualization code
Noted during visualization development that using hex code version of
the image hash for the raw image uuid in the DB was messing with the
ability to embed that data in a link. Redid to use a int representation
of the hash instead.

Build a basic flask visualization with the various html pages resource
servers (images taken from the DB) and post activities (relabeling
images) anticipated for the final application. Still needs lots of
visual TLC. Uses an nearly identical set of config params to the
InitInputService